+++ app/models/Order.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class Order(id: Int, orderDate: String, userId: Int, paymentId: Int)
+case class Order(id: Long, orderDate: String, userId: Long, paymentId: Long)
 
 object Order {
   implicit val orderJSON = Json.format[Order]
